requirements.txt
Pillow==11.2.1
qrcode==8.1
